* changes:
  Fix some more of size_t and unsigned ints in DALi API
  Reduce the binary size of Constraint code by refactoring the templates to produce less bloat
  Refactored Animator classes to reduce code binary size by 60%

+++ dali/internal/event/common/type-info-impl.h
@@ -93,22 +93,22 @@ public:
   /**
    * @copydoc Dali::TypeInfo::GetActionCount
    */
-  size_t GetActionCount() const;
+  uint32_t GetActionCount() const;
 
   /**
    * @copydoc Dali::TypeInfo::GetActionName
    */
-  std::string GetActionName(size_t index) const;
+  std::string GetActionName( uint32_t index ) const;
 
   /**
    * @copydoc Dali::TypeInfo::GetSignalCount
    */
-  size_t GetSignalCount() const;
+  uint32_t GetSignalCount() const;
 
   /**
    * @copydoc Dali::TypeInfo::GetSignalName
    */
-  std::string GetSignalName(size_t index) const;
+  std::string GetSignalName( uint32_t index ) const;
 
   /**
    * @copydoc Dali::TypeInfo::GetPropertyCount
